+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this package are documented in this file. This project adheres to [Haskell PVP](https://pvp.haskell.org/) versioning.
 
-## Unreleased
+## 0.4.0.0
 
 - #82, Implement `search()` function from function extensions
 - #83, Improve parse errors